refactor(cli): move ts files into src and clean up prerender
