+++ cli/src/static.paths.ts
@@ -1,4 +1,4 @@
-module.exports = [
+export const ROUTES = [
   '/',
   '/lazy',
   '/lazy/nested'
